ML Models/test_db.py
import unittest
from unittest.mock import patch, MagicMock
import pandas as pd
import numpy as np

# Import the necessary functions from your main script
from db import train_model, fetch_sensor_data, suggest_maintenance_tasks, store_predictions, main

class TestMotorFailurePrediction(unittest.TestCase):

    def setUp(self):
        # Set up any required initializations, such as mock data
        self.mock_engine = MagicMock()
        self.mock_conn = MagicMock()
        self.mock_engine.connect.return_value.__enter__.return_value = self.mock_conn

        self.historical_data = pd.DataFrame({
            'temperature': [70, 75, 80, 85, 90],
            'vibration': [0.2, 0.3, 0.4, 0.5, 0.6],
            'current': [10, 12, 14, 16, 18],
            'failure': [0, 0, 1, 1, 1]
        })

        self.real_time_data = pd.DataFrame({
            'timestamp': pd.to_datetime(['2024-08-24 10:15:38', '2024-08-24 10:16:38']),
            'motor_id': [1, 2],
            'temperature': [88, 92],
            'vibration': [0.55, 0.65],
            'current': [17, 19]
        })

    @patch('db.fetch_historical_data')
    def test_train_model_no_data(self, mock_fetch_historical_data):
        # Mock fetching historical data to return an empty DataFrame
        mock_fetch_historical_data.return_value = pd.DataFrame()

        model = train_model()
        self.assertIsNone(model, "Model should be None when no historical data is available.")


    def test_suggest_maintenance_tasks(self):
        # Test the maintenance task suggestions based on different failure probabilities
        high_risk_task = suggest_maintenance_tasks(0.9)
        self.assertEqual(high_risk_task, "Immediate action required: Inspect motor, replace bearings, lubricate, and run full diagnostics.")

        medium_risk_task = suggest_maintenance_tasks(0.7)
        self.assertEqual(medium_risk_task, "High priority: Perform detailed inspection, check for overheating, and lubricate motor.")

        low_risk_task = suggest_maintenance_tasks(0.1)
        self.assertEqual(low_risk_task, "Low priority: Routine check-up and basic maintenance.")

    @patch('db.store_predictions')
    def test_store_predictions(self, mock_store_predictions):
        # Test if storing predictions works correctly
        mock_store_predictions(self.mock_engine, [
            ('2024-08-24 10:15:38', 1, 0.95, True, "Immediate action required: Inspect motor, replace bearings, lubricate, and run full diagnostics.")
        ])

        self.assertTrue(mock_store_predictions.called, "store_predictions should be called with data.")

    def tearDown(self):
        # Any cleanup actions, if necessary
        pass

if __name__ == '__main__':
    unittest.main()
